requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==18.2.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dnspython==2.6.1
doc2docx==0.2.4
doc2pdf==0.2.0
docx2pdf==0.1.8
easyocr==1.7.1
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.113.0
fastapi-cli==0.0.5
filelock==3.15.4
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
imageio==2.35.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
langdetect==1.0.9
lazy_loader==0.4
lit==18.1.8
lmdb==1.5.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.23.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandoc==2.4
pdflatex==0.1.3
pdfminer==20191125
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
plumbum==1.8.3
ply==3.11
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.0
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.23.2
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypandoc==1.13
pyparsing==3.1.4
PyPDF2==3.0.1
pypdfium2==4.30.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
reportlab==4.2.2
requests==2.32.3
rich==13.8.0
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
sympy==1.13.2
tifffile==2024.8.30
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.5
triton==2.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
